requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.119.1
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1

tortoise-orm==0.24.2
asyncpg==0.30.0
bcrypt==4.3.0
PyJWT==2.10.1
python-multipart==0.0.20